+++ nbed/exceptions.py
@@ -1,12 +1,6 @@
 """Custom Exceptions."""
 
 
-class NbedConfigError(Exception):
-    """Raised when config is not valid."""
-
-    pass
-
-
 class NbedDriverError(Exception):
     """Raise when NbedDriver finds itself in a bad state."""
 
@@ -15,12 +9,6 @@ class NbedDriverError(Exception):
 
 class NbedLocalizerError(Exception):
     """Raise when Localizer sense check fails."""
-
-    pass
-
-
-class HamiltonianConverterError(Exception):
-    """Base Exception class."""
 
     pass