+++ nbed/utils.py
@@ -1,20 +1,17 @@
 """Helper functions for the package."""
 
 import argparse
+import json
 import logging
 import logging.config
 import os
 from pathlib import Path
 from typing import Optional
 
-import yaml
-from openfermion import count_qubits
 from openfermion.chem.pubchem import geometry_from_pubchem
+from pydantic import ValidationError
 
-from nbed.ham_builder import HamiltonianBuilder
-
-from .config import Projector
-from .driver import NbedDriver
+from nbed.config import NbedConfig
 
 logger = logging.getLogger(__name__)
 
@@ -52,338 +49,32 @@ def setup_logs() -> None:
     logger.debug("Logging initialised.")
 
 
-def restricted_float_percentage(x: float) -> float:
-    """Checks input x is within 0-1 range (percentage) and is a float.
-
-    Args:
-        x (float): input number between 0 and 1 (inclusive)
-
-    Returns:
-        x (float): input percentage
-    """
-    try:
-        x = float(x)
-    except ValueError:
-        raise argparse.ArgumentTypeError("{!r} not a floating-point literal".format(x))
-
-    if x < 0.0 or x > 1.0:
-        raise argparse.ArgumentTypeError("{!r} not in range [0.0, 1.0]".format(x))
-    return x
-
-
 def parse():
     """Parse arguments from command line interface."""
     logger.debug("Adding CLI arguments.")
     parser = argparse.ArgumentParser(description="Output embedded Qubit Hamiltonian.")
     parser.add_argument(
         "--config",
+        required=True,
         type=str,
         help="Path to a config file. Overwrites other arguments.",
-    )
-    parser.add_argument(
-        "--geometry",
-        "-g",
-        type=str,
-        help="Path to an XYZ file or raw xyz string of molecular structure (note active atoms must appear first).",
-    )
-    parser.add_argument(
-        "--active_atoms",
-        "-a",
-        type=int,
-        help="Number of atoms to include in active region.",
-    )
-    parser.add_argument(
-        "--basis",
-        "-b",
-        type=str,
-        help="Basis set to use.",
-    )
-    parser.add_argument(
-        "--xc_functional",
-        "--xc",
-        "-x",
-        type=str,
-        help="Exchange correlation functional to use in DFT calculations.",
-    )
-    parser.add_argument(
-        "--projector",
-        "-p",
-        type=str,
-        choices=["huzinaga", "mu"],
-        help="Which projector method to use.",
-    )
-    parser.add_argument(
-        "--unit",
-        "-u",
-        type=str,
-        choices=["angstrom", "bohr"],
-        help="Distance unit of molecular geometry",
-    )
-    parser.add_argument(
-        "--localization",
-        "--loc",
-        "-l",
-        type=str.lower,
-        choices=["spade", "pipek-mezey", "ibo", "boys"],
-        help="Method of localization to use.",
-    )
-    parser.add_argument(
-        "--output",
-        "-o",
-        type=str.lower,
-        choices=["openfermion", "qiskit", "pennylane"],
-        help="Quantum computing backend to output the qubit hamiltonian for.",
-    )
-    parser.add_argument(
-        "--convergence",
-        "--conv",
-        "-c",
-        type=float,
-        help="Convergence tolerance for calculations.",
-    )
-    parser.add_argument(
-        "--charge",
-        type=int,
-        help="Charge of molecular system.",
-    )
-    parser.add_argument(
-        "--spin",
-        type=int,
-        help="Spin of molecular system.",
-    )
-    parser.add_argument(
-        "--savefile",
-        "-s",
-        type=str,
-        help="Path to save file.",
-    )
-    parser.add_argument(
-        "--run_ccsd_emb",
-        action="store_true",
-        help="Include if you want to run a ccsd calculation of the active embedded system.",
-    )
-    parser.add_argument(
-        "--run_fci_emb",
-        action="store_true",
-        help="Include if you want to run a fci calculation of the active embedded system.",
-    )
-    parser.add_argument(
-        "--ram",
-        type=str,
-        help="amount of ram in MB that PySCF can use",
-    )
-    parser.add_argument(
-        "--mu_shift",
-        type=int,
-        help="mu energy shift value",
-    )
-    parser.add_argument(
-        "--virtual_localization",
-        "--virt_loc",
-        action="store_true",
-        help="whether to run localization of virutal (unoccupied) orbitals",
-    )
-    parser.add_argument(
-        "--occupied_threshold",
-        "--ot",
-        type=restricted_float_percentage,
-        help="occupation threshold (float between 0 and 1 inclusive) used to localize occupied molecular orbs (unnecessary for spade approach)",
-    )
-    parser.add_argument(
-        "--virtual_threshold",
-        "--vt",
-        type=restricted_float_percentage,
-        help="threshold (float between 0 and 1 inclusive) used to localize unoccupied (virtual) molecular orbs (necessary for spade approach)",
-    )
-    parser.add_argument(
-        "--max_hf_cycles",
-        "--hf_c",
-        type=int,
-        help="max number of Hartree-Fock iterations allowed for global and local Hartree-Fock calcs",
-    )
-    parser.add_argument(
-        "--max_dft_cycles",
-        "--dft_c",
-        type=int,
-        help="max number of DFT iterations allowed in scf calc",
     )
     logger.debug("Parsing CLI arguments.")
     args = parser.parse_args()
 
-    if args.config:
-        logger.debug("Reading config file.")
-        filepath = Path(args.config).absolute()
-        stream = open(filepath)
-        args = yaml.safe_load(stream)["nbed"]
+    logger.debug("Reading config file.")
+    filepath = Path(args.config).absolute()
+    with open(filepath) as f:
+        config_data = json.load(f)
+    logger.debug(f"Input data:\n{config_data=}")
 
-        # Optional argument defaults
-        args["unit"] = args.get("unit", "angstrom")
-        args["charge"] = args.get("charge", 0)
-        args["spin"] = args.get("spin", 0)
-        args["convergence"] = args.get("convergence", 1e-6)
-        args["run_ccsd_emb"] = args.get("run_ccsd_emb", False)
-        args["run_fci_emb"] = args.get("run_fci_emb", False)
-        args["mu_shift"] = args.get("mu_shift", 1e6)
-        args["ram"] = args.get("ram", 4_000)
-        args["virtual_localization"] = args.get("virtual_localization", False)
-        args["occupied_threshold"] = args.get("occupied_threshold", 0.95)
-        args["virtual_threshold"] = args.get("virtual_threshold", 0.95)
-        args["max_hf_cycles"] = args.get("max_hf_cycles", 50)
-        args["max_dft_cycles"] = args.get("max_dft_cycles", 50)
-    else:
-        # Transform the namespace object to a dict.
-        args = vars(args)
+    try:
+        config = NbedConfig(config_data)
+    except ValidationError as e:
+        logger.error("Could not validate input data against NbedConfig model.")
+        logger.error(e)
 
-    if any([values is None for values in args.values()]):
-        logger.info(
-            f"Missing values for argument {[key for key, value in args.items() if value is None]}"
-        )
-        logger.info("\nMissing values for arguments: ".upper())
-        logger.info(f"{[key for key, value in args.items() if value is None]}\n")
-        raise Exception("Missing argument values.")
-
-    args["savefile"] = str(Path(args["savefile"]).absolute())
-    args["convergence"] = float(args["convergence"])
-
-    logger.debug(f"Arguments: {args}")
-    return args
-
-
-def print_summary(driver: NbedDriver, transform: str, fci: bool = False) -> None:
-    """Print a summary of the package results.
-
-    Args:
-        driver (NbedDriver): An NbedDriver to summarise.
-        transform (str): The transform used to generate a qubit Hamiltonian.
-        fci (bool): Whether to run full system fci.
-    """
-    logger.debug("Printing summary of results.")
-    # for get statements
-    default = "Not calculated."
-
-    qham = HamiltonianBuilder(
-        driver.embedded_scf,
-        constant_e_shift=driver.classical_energy,
-        transform=transform,
-    ).build()
-
-    # Would be a great place for a switch statemet when
-    # dependencies catch up with python 3.10
-    match driver.config.projector:
-        case Projector.BOTH:
-            mu_qham, huz_qham = qham
-        case Projector.HUZ:
-            mu_qham, huz_qham = None, qham
-        case Projector.MU:
-            mu_qham, huz_qham = qham, None
-
-    print("".center(80, "*"))
-    logger.info("".center(80, "*"))
-    print("  Summary of Embedded Calculation".center(80))
-    logger.info("  Summary of Embedded Calculation".center(80))
-    print("".center(80, "*"))
-    logger.info("".center(80, "*"))
-
-    print(f"global (cheap) DFT calculation {driver._global_ks.e_tot}")
-    logger.info(f"global (cheap) DFT calculation {driver._global_ks.e_tot}")
-
-    if driver.config.projector in [Projector.HUZ, Projector.BOTH]:
-        print("".center(80, "*"))
-        logger.info("".center(80, "*"))
-        print("  Huzinaga calculation".center(20))
-        logger.info("  Huzinaga calculation".center(20))
-        print(
-            f"Total energy - active system at RHF level: {driver._huzinaga.get('e_rhf', default)}"
-        )
-        logger.info(
-            f"Total energy - active system at RHF level: {driver._huzinaga.get('e_rhf', default)}"
-        )
-        if driver.config.run_ccsd_emb is True:
-            print(
-                f"Total energy - active system at CCSD level: {driver._huzinaga.get('e_ccsd', default)}"
-            )
-            logger.info(
-                f"Total energy - active system at CCSD level: {driver._huzinaga.get('e_ccsd', default)}"
-            )
-        if driver.config.run_fci_emb is True:
-            print(
-                f"Total energy - active system at FCI level: {driver._huzinaga.get('e_fci', default)}"
-            )
-            logger.info(
-                f"Total energy - active system at FCI level: {driver._huzinaga.get('e_fci', default)}"
-            )
-
-        print(
-            f"length of huzinaga embedded fermionic Hamiltonian: {len(huz_qham.terms)}"
-        )
-        logger.info(
-            f"length of huzinaga embedded fermionic Hamiltonian: {len(huz_qham.terms)}"
-        )
-        print(f"number of qubits required: {count_qubits(huz_qham)}")
-        logger.info(f"number of qubits required: {count_qubits(huz_qham)}")
-
-    if driver.config.projector in [Projector.MU, Projector.BOTH]:
-        print("".center(80, "*"))
-        logger.info("".center(80, "*"))
-        print("  Mu shift calculation".center(20))
-        logger.info("  Mu shift calculation".center(20))
-        print(
-            f"Total energy - active system at RHF level: {driver._mu.get('e_rhf', default)}"
-        )
-        logger.info(
-            f"Total energy - active system at RHF level: {driver._mu.get('e_rhf', default)}"
-        )
-        if driver.config.run_ccsd_emb is True:
-            print(
-                f"Total energy - active system at CCSD level: {driver._mu.get('e_ccsd', default)}"
-            )
-            logger.info(
-                f"Total energy - active system at CCSD level: {driver._mu.get('e_ccsd', default)}"
-            )
-        if driver.config.run_fci_emb is True:
-            print(
-                f"Total energy - active system at FCI level: {driver._mu.get('e_fci', default)}"
-            )
-            logger.info(
-                f"Total energy - active system at FCI level: {driver._mu.get('e_fci', default)}"
-            )
-
-        print(f"length of mu embedded fermionic Hamiltonian: {len(mu_qham.terms)}")
-        logger.info(
-            f"length of mu embedded fermionic Hamiltonian: {len(mu_qham.terms)}"
-        )
-        print(f"number of qubits required: {count_qubits(mu_qham)}")
-        logger.info(f"number of qubits required: {count_qubits(mu_qham)}")
-
-    logger.debug("Building full system Hamiltonian for comparison.")
-    full_system_hamiltonian = HamiltonianBuilder(
-        driver._global_hf, constant_e_shift=0, transform=transform
-    ).build()
-
-    print("".center(80, "*"))
-    logger.info("".center(80, "*"))
-    print("  Summary of reference Calculation".center(80))
-    logger.info("  Summary of reference Calculation".center(80))
-    print("".center(80, "*"))
-    logger.info("".center(80, "*"))
-
-    if fci:
-        print("Running Full system FCI and preparing Hamiltonian.")
-        logger.info("Running Full system FCI and preparing Hamiltonian.")
-        print(f"Global (expensive) full FCI calculation {driver._global_fci.e_tot}")
-        logger.info(
-            f"Global (expensive) full FCI calculation {driver._global_fci.e_tot}"
-        )
-
-    print(
-        f"length of full system fermionic Hamiltonian: {len(full_system_hamiltonian.terms)}"
-    )
-
-    logger.info(
-        f"length of full system fermionic Hamiltonian: {len(full_system_hamiltonian.terms)}"
-    )
-    print(f"number of qubits required: {count_qubits(full_system_hamiltonian)}")
-    logger.info(f"number of qubits required: {count_qubits(full_system_hamiltonian)}")
+    return config
 
 
 def pubchem_mol_geometry(molecule_name) -> dict: